requirements.txt

pytz==2024.1



requests==2.32.3
joblib==1.4.2




# Core
streamlit==1.33.0
pandas==2.0.3
numpy==1.26.4

# Network/Threat Analysis
scapy==2.5.0
dnspython==2.4.2
geoip2==4.8.0
python-whois==0.9.0
python-nmap==0.7.1
virustotal-api==1.1.11
python-dateutil==1.17.0
# Visualization
folium==0.14.0 
streamlit-folium==0.15.1
matplotlib==3.7.1

# PDF Reporting
reportlab==4.0.4

# Security/SSL
pyOpenSSL==23.3.0

# Utilities
ipaddress==1.0.23
logging==0.4.9.6
python-dotenv==1.0.0
